requirements.txt
black
flake8
isort
-e .
requests==2.28.2
python-dotenv==1.0.0
openai==1.38.0
pymongo==4.8.0